Fix bug after conflict
